A lot of more comments and eliminating utils.
