+++ internal/photoprism/importer.go
@@ -2,6 +2,7 @@ package photoprism
 
 import (
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"path"
@@ -154,4 +155,22 @@ func (i *Importer) GetDestinationFilename(mainFile *MediaFile, mediaFile *MediaF
 	}
 
 	return result, nil
+}
+
+func directoryIsEmpty(path string) bool {
+	f, err := os.Open(path)
+
+	if err != nil {
+		return false
+	}
+
+	defer f.Close()
+
+	_, err = f.Readdirnames(1)
+
+	if err == io.EOF {
+		return true
+	}
+
+	return false
 }
