+++ internal/photoprism/mediafiles.go
@@ -1,22 +1,26 @@
 package photoprism
 
+// MediaFiles provides a Collection for mediafiles.
 type MediaFiles []*MediaFile
 
+// Len returns the length of the mediafile collection.
 func (f MediaFiles) Len() int {
 	return len(f)
 }
 
+// Less compares two mediafiles based on filename length.
 func (f MediaFiles) Less(i, j int) bool {
 	fileName1 := f[i].GetFilename()
 	fileName2 := f[j].GetFilename()
 
 	if len(fileName1) == len(fileName2) {
 		return fileName1 < fileName2
-	} else {
-		return len(fileName1) < len(fileName2)
 	}
+
+	return len(fileName1) < len(fileName2)
 }
 
+// Swap changes two files order around.
 func (f MediaFiles) Swap(i, j int) {
 	f[i], f[j] = f[j], f[i]
 }
