+++ internal/photoprism/thumbnails_test.go
@@ -1,8 +1,9 @@
 package photoprism
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestMediaFile_GetThumbnail(t *testing.T) {
